parser: Fix crash on non-ASCII char literals
If the given char literal was not an ASCII one, there was the chance for
the character iterator to mess things up. Hence, when checking the
closing single quote, it might encounter a None value.

This is simply mitigated my moving the check of ASCII alphanumeric
before checking for the closing quote.

Fixes: b1623996c76f ("Add support for character literals")

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>