Improve the format of error messages
It was weird to show warnings which also showed "Error: " as a message,
and likewise it was weird for errors to display their kind, since users
simply do not care about this kind of information.

Hence, streamline the format to something closer to what it's done by
modern assemblers/compilers.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>