Shrink some absolute instructions by one byte
Some instructions which make use of absolute addressing can actually be
further compressed to act like zeropage indexing. This can be done if
the immediate being used by that instruction can actually fit into a
single byte. If that's the case, we will then "correct" the programmer
by using zeropage indexing instead of an absolute one, thus reducing one
byte for that instruction.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>