Dual license the source code under this repository
This means that the source code under `crates/` is under the GNU
GPLv3 (or any later version), and that source code under `lib/` is under
the GNU LGPLv3 (or any later version).

In practice this relaxes a bit the previous licensing because before it
was all GNU GPLv3 (or any later version). For people compiling a binary
with these libraries then this would've been spilling over the GPLv3 on
their binaries as well, which might not be desired. Instead, for the
libraries abide by the LGPLv3 (or any later version), which has the same
guarantees when it comes to free/lliure software for this specific case,
but at least we don't force the GPLv3 (or any later version) onto users.
That being said, users of these libraries still need to provide object
files so third parties could theoretically recompile those binaries
under a modified library.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>