Introduce end-to-end tests
For now I have added the most simple example for an NES/Famicom game,
which I already had on a private repository (that I plan to open source
soonish).

Tests simply run `nasm` on known source files and compares the results
with an .nes ROM file that has the exact bytes that we expect.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>