parser: Implement block bodies
These are bodies which are the right node of some proc controls. This
way the parser comes back to sanity for statements like .macro and the
likes, and behaves more like a usual parser. This was not done in the
past because I thought things could have been simpler this way, but it
ended up making the assembler way more complicated that it needed to.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>