Try absolute addressing on bad zeropage addressing
Sometimes, out of clarity, the programmer may have written something
along the lines of:

    lda $40, y

This is invalid because the `lda` instruction does not allow zeropage
y-indexing addressing mode. That being said, it does allow for absolute
y-indexing addressing mode. This commit allows this syntax by
transforming code like the previous one into:

    lda $0040, y

This cannot always be done, but the assembler should at least try if
it's possible and not trouble the programmer.

That being said, this is otherwise a bit shady since the programmer
might think that it's a 2-byte instruction when it's a 3-byte one.
Hence, maybe a future linter can pick up code like this and suggest a
more explicit writing.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>