Restrict literal characters further
In bad314e1cfbf ("Prevent numeric literals from having spaces") it was
added already the restriction on not having whitespace characters in
literal expressions. Here we go a step further and we more explicitely
limit which symbol combinations can go into a literal declaration (e.g.
"#$2" is valid but "##2" is not).

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>