Evaluate bare numbers as decimal values
In places like macro calls, the programmer might actually prefer to pass
a numeric argument as is, without any prefixes. In these cases, just
evaluate it as a decimal. In fact, this was already covered when
evaluating the context because the same thing happens to assignments.
Hence, I just needed to expand the scope of this.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>