Prevent assignments on relative references
It is a really weird thing to do, but on a twisted way I can see how
someone could think of such a monstrosity. Panicking on such a case is
not valid because that's not the fault from the assembler but from the
programmer, and so a proper message should be displayed instead.

This was detected via the fuzzer, but it was a side effect from
e7c5e63d04f6 ("Evaluate bare numbers as decimal values").

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>