Make errors from nasm itself more cohesive
The errors reported from the library had a slightly different format
than those that happened in the main file for nasm. This is because in
these cases we were just returning an Err type and Rust handles that
through a special impl.

Instead of going the route of type masturbation, let's go the C route
and have a 'die' function that simply prints the message and calls
exit(1),

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>