Implement the .res control statement
Similarly to other assemblers, this allows the programmer to write a
definite amount of bytes with the same values.

Compared to other assemblers there are two things to notice. First,
there is a limit to it (i.e. whatever can fit in 2 bytes). Second, if
the fill value is not provided, then it will default to the current
mapping's fill value, or just 0x00 if the current mapping doesn't define
one of its own.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>