Parse a string literal as a single object
Before this it was left to the `parse_identifier` to figure things out,
but this was prone to silly errors like "a: b", in which it would
mistake it as the start of a label. Instead of any of this, just consume
a string literal if it has been detected.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>