Allow semicolons inside of strings
The parser was too naive and assumed that a semicolon immediately
implied an inline comment, and that might not just be the case as it is
with string literals.

Hence, the end of each semantic line has to consider not only whether
there is an inline comment, but also if a string literal is being used
and whether it surrounds or not the given semicolon.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>