Fix error with strings containing equal operator
As with e27a1593c8d8 ("Allow semicolons inside of strings"), the parser
was too naive and regarded any '=' operator as part of an assignment,
despite that it could be art of string literal.

Luckily the fix was already done inside of the parsing of assignments,
we just needed to move it up.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>