parser: Fix disambiguation paren on first argument
Arguments can be put inside of enclosing parenthesis, but the parser was
assuming that if an opening paren was found when parsing the first
argument on an argument list, then that was all it was needed to be
parsed. This though conflicts with situations like:

       .byte ($01 << 2) | ($01 << 1)

In this case, the parser would have ignored everything past the first
closing paren.

This commit provides a fix in which if an operation is found past the
first enclosing parenthesis, then this assumption is discarded in favor
of a parenthesis being used for disambiguating on an
arithmetical/logical expression.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>