nasm: Error out when using WRAM when not available
This commit introduces the ability to inspect the temptative header
before producing the actual output, and with that it checks whether the
Working RAM is being advertised or not. If it is not being advertised
but the assembler detected memory accesses to that region, then we are
in trouble and we should error out.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>