xixanta: Don't allow negative variables on asan
Negative variables make no sense when it comes to do stuff like:

         variable = -1
         lda variable

That being said, as an assembler you never know the hacks and nonsense
programmers are willing to endure. But we do know that if the address
sanitizer is enabled, since if it's an "asan-friendly", then an explicit
negative variable can be barred.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>