Report on multi-byte values on arithmetic evaluation
When evaluating an arithmetic operation, check on whether it
theoretically spans more than one byte, and set it on the resulting
Bundle.size. It's up to the caller to decide on whether that makes sense
or not. This fixes situations in which:

   sta $200 + 1

was marked as having 2 bytes for the size as only the "1" was being
considered for the size on this operation. This would then signal the
"sta" evaluation that it was zeropage instead of absolute addressing.

Signed-off-by: Miquel Sabaté Solà <mssola@mssola.com>